mwaa_aws_managed_airflow/requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-1.10.12/constraints-3.7.txt"
apache-airflow-backport-providers-amazon
smart_open